requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-dotenv==1.0.0
google-cloud-aiplatform==1.38.1
requests==2.31.0
aiohttp==3.9.1
setuptools>=65.0.0
yfinance==0.2.28
python-dateutil==2.8.2